requirements.txt
pytest
selenium
webdriver-manager
PyHamcrest
behave